--- a/common/changes/@artusx/plugin-koa/feature-find-my-way_2024-03-02-07-57.json
+++ b/common/changes/@artusx/plugin-koa/feature-find-my-way_2024-03-02-07-57.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@artusx/plugin-koa",
+      "comment": "feat: use find-my-way to improve perf",
+      "type": "none"
+    }
+  ],
+  "packageName": "@artusx/plugin-koa"
+}