--- a/packages/apps/artusx-koa-bench/.eslintignore
+++ b/packages/apps/artusx-koa-bench/.eslintignore
@@ -1,0 +1,4 @@
+node_modules
+lib
+dist
+coverage