--- a/packages/apps/artusx-koa-bench/.eslintrc
+++ b/packages/apps/artusx-koa-bench/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": ["@artusx/eslint-config"],
+  "parserOptions": {
+    "project": "./tsconfig.json"
+  }
+}