--- a/packages/apps/artusx-koa-bench/README.md
+++ b/packages/apps/artusx-koa-bench/README.md
@@ -1,0 +1,16 @@
+# artusx-koa-bench
+
+> benchmark for @artusx/core
+
+## Usage
+
+```
+rush install && rush build && pnpm bench:fmy
+
+```
+
+
+### Requirement
+
+- Node.js 18.x
+- wrk