--- a/packages/apps/artusx-koa-bench/src/config/plugin.development.ts
+++ b/packages/apps/artusx-koa-bench/src/config/plugin.development.ts
@@ -1,0 +1,11 @@
+export default {
+  redis: {
+    enable: false,
+  },
+  sequelize: {
+    enable: false,
+  },
+  // xtransit: {
+  //   enable: true,
+  // },
+};