--- a/packages/apps/artusx-koa-bench/src/config/plugin.ts
+++ b/packages/apps/artusx-koa-bench/src/config/plugin.ts
@@ -1,0 +1,6 @@
+export default {
+  artusx: {
+    enable: true,
+    package: '@artusx/core',
+  },
+};