--- a/packages/apps/artusx-koa-bench/src/constants/index.ts
+++ b/packages/apps/artusx-koa-bench/src/constants/index.ts
@@ -1,0 +1,1 @@
+export const name = 'artusx-bench';