--- a/packages/apps/artusx-koa-bench/src/controller/home.ts
+++ b/packages/apps/artusx-koa-bench/src/controller/home.ts
@@ -1,0 +1,11 @@
+import { Controller, GET } from '@artusx/core';
+
+import type { ArtusxContext } from '@artusx/core';
+
+@Controller()
+export default class HomeController {
+  @GET('/')
+  async home(ctx: ArtusxContext) {
+    ctx.body = {};
+  }
+}