--- a/packages/apps/artusx-koa-bench/src/index.ts
+++ b/packages/apps/artusx-koa-bench/src/index.ts
@@ -1,0 +1,3 @@
+import { main } from './bootstrap';
+
+main();