--- a/packages/apps/artusx-koa-bench/tsconfig.json
+++ b/packages/apps/artusx-koa-bench/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@artusx/tsconfig",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "outDir": "dist"
+  },
+  "include": ["src/**/*.ts"]
+}