--- a/packages/plugins/koa/package.json
+++ b/packages/plugins/koa/package.json
@@ -53,7 +53,7 @@
   "dependencies": {
     "@artus/core": "^2.x",
     "@artus/pipeline": "^0.2",
-    "@koa/router": "^12.0.1",
+    "find-my-way": "^8.1.0",
     "koa": "^2.15.0"
   },
   "devDependencies": {
@@ -61,7 +61,6 @@
     "@artusx/tsconfig": "workspace:*",
     "@types/jest": "~29.5.11",
     "@types/koa": "^2.13.12",
-    "@types/koa__router": "^12.0.4",
     "@types/node": "^18.11.17",
     "@typescript-eslint/eslint-plugin": "~6.19.1",
     "eslint": "~8.56.0",