To benchmark in `main` branch, I disabled both `checkAuth` and `LimitRate` middlewares. Use `FACTOR` multiply `11` to gen multiple routes.

The `Requests/sec` result compare:

| FACTOR| @koa/router | find-my-way |
| -------- | ------- | -------- |
| 1 | 11633.29 | 14644.27 |
| 5 | 10474.16 | 12408.07 |
| 10 | 9823.64 | 13710.47 |
| 20 | 5909.84 | 9150.35 |
| 50 | 4001.26 | 3552.96 |
| 100 | 2554.90 | 5645.61 |
| 200 | 1343.33 | 9383.24 |
| 500 | 540.95 | 9198.66 |
| 1000 | 279.51 | 8927.13 |


`factor=5` details

@koa/router
```
Running 3s test @ http://localhost:7001/10/child/grandchild/%40
  8 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.51ms    1.65ms  27.54ms   88.32%
    Req/Sec     1.35k   639.23    10.32k    96.27%
  32471 requests in 3.10s, 8.98MB read
  Non-2xx or 3xx responses: 32471
Requests/sec:  10474.16
Transfer/sec:      2.90MB
```

find-my-way
```
Running 3s test @ http://localhost:7001/10/child/grandchild/%40
  8 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.92ms    1.43ms  23.40ms   89.97%
    Req/Sec     1.56k   315.52     1.98k    74.60%
  38523 requests in 3.10s, 8.82MB read
  Non-2xx or 3xx responses: 38523
Requests/sec:  12408.07
Transfer/sec:      2.84MB
```